pyproject.toml
[tool.poetry]
name = "hw-lms"
version = "0.1.0"
description = ""
authors = ["Vadim Grokhovskiy <grokhovskiy.job@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
